requirements.txt
streamlit>=1.44.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.1
openai>=1.51.0
chromadb>=0.5.5
pypdf>=5.0.1
pydantic>=2.9.2
python-multipart>=0.0.9
